src/Engine/pyproject.toml
[project]
name = "streamstorm"
version = "3.5.2"
description = "Spam chats on YouTube live streams"
requires-python = ">=3.13"
dependencies = [
    "psutil>=6.1.0",
    "pydantic>=2.10.4",
    "selenium>=4.27.1",
    "platformdirs>=4.3.8",
    "undetected-chromedriver>=3.5.5",
    "setuptools>=80.9.0",
    "pywebview[qt]>=5.4 ; sys_platform == 'linux'",
    "pywebview>=5.4 ; sys_platform != 'linux'",
    "fastapi[standard]>=0.115.14",
    "uvicorn>=0.40.0",
    "playwright>=1.53.0",
    "aiofiles>=24.1.0",
    "dgupdater>=1.2.1",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-community>=0.4.1",
    "pyautogui>=0.9.54",
    "python-socketio>=5.16.0",
    "yt-dlp>=2025.11.12",
    "pydantic-ai-slim[anthropic,fastmcp,google,openai]>=0.4.3",
    "pydantic-settings>=2.12.0",
    "fastmcp>=2.14.1",
]

[dependency-groups]
dev = [
    "auto-py-to-exe>=2.46.0",
    "pyinstaller>=6.15.0",
    "tomlkit>=0.13.3",
    "atatus>=2.2.0",
    "python-dotenv>=1.1.1",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]
